% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfisher.R
\name{vfisher.test}
\alias{vfisher.test}
\title{Vectorized fisher.test}
\usage{
vfisher.test(
  a,
  b,
  c,
  d,
  conf.int = TRUE,
  conf.level = 0.95,
  or = rep(1, length(a)),
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{a}{\code{x[1, 1]} in the matrix version of \code{\link[stats:fisher.test]{stats::fisher.test()}}}

\item{b}{\code{x[1, 2]} in the matrix version of \code{\link[stats:fisher.test]{stats::fisher.test()}}}

\item{c}{\code{x[2, 1]} in the matrix version of \code{\link[stats:fisher.test]{stats::fisher.test()}}}

\item{d}{\code{x[2, 2]} in the matrix version of \code{\link[stats:fisher.test]{stats::fisher.test()}}}

\item{conf.int}{logical indicating if a confidence interval for the
    odds ratio in a \eqn{2 \times 2}{2 by 2} table should be
    computed (and returned).}

\item{conf.level}{confidence level for the returned confidence
    interval.  Only used in the \eqn{2 \times 2}{2 by 2} case and if
    \code{conf.int = TRUE}.}

\item{or}{the hypothesized odds ratio.  Only used in the
    \eqn{2 \times 2}{2 by 2} case.}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.
    You can specify just the initial letter.  Only used in the
    \eqn{2 \times 2}{2 by 2} case.}
}
\value{
data.table, columns \code{a}, \code{b}, \code{c}, \code{d}, \code{or}, \code{p.value}, \code{estimate}.

If \code{conf.int == TRUE} (the default), will also include columns for the
confidence interval, \code{ci.lo} and \code{ci.hi}. If \code{alternative == "less"}, \code{ci.lo}
will be \code{0}, and similarly for \code{alternative == "greater"}, \code{ci.hi} will be
\code{Inf}. Otherwise (the default), the CI will be centered and both low and high
ends will take on values between 0 and Inf.
}
\description{
Provides a vectorized version of \code{\link[stats:fisher.test]{stats::fisher.test()}} for evaluating a
series of 2x2 contingency tables.
}
\details{
N.b. \code{vfisher.test} does less input validation than \code{\link[stats:fisher.test]{stats::fisher.test()}}.

Because \code{vfisher.test} does not support anything other than 2x2 tests, the
arguments to \code{fisher.test} associated with other tests (e.g. \code{hybrid}) do not
appear.
}
